update
Q1 先考虑无噪环境下如何逼近fci：
1. 增加一些可学习参数线路确实比纯HF精度高，且参数不用太多（甚至连ry_HAE都多了）
2. 对于H4问题来说最优解的参数非常接近pi的分数，这意味着结构可以写死；且如果扇出了RY门，则甚至可以省略HF线路
3. optimizer稳定性问题，COBYLA和BFGS都有大概率陷入局部最优
Q3 尝试实现LM-BS但失败

+++ stage1_Q3#QAIA/vis_data.py
@@ -15,6 +15,9 @@ BASE_PATH = Path(__file__).parent
 DATA_PATH = BASE_PATH / 'MLD_data'
 IMG_PATH = BASE_PATH / 'img' ; IMG_PATH.mkdir(exist_ok=True)
 
+# https://stackoverflow.com/questions/15505514/binary-numpy-array-to-list-of-integers
+bits_to_number = lambda bits: bits.dot(1 << np.arange(bits.shape[-1] - 1, -1, -1))
+
 
 @dataclass
 class Sample:
@@ -30,7 +33,6 @@ class Sample:
   @property
   def x(self) -> ndarray:
     # x ~= H' * y (?)
-    breakpoint()
     return np.linalg.inv(self.H) @ self.y
 
 
@@ -41,6 +43,11 @@ def get_dataset() -> List[Sample]:
 
 
 def plot_hist(features:List[ndarray], title:str='feat'):
+  fp = IMG_PATH / f'{title}.png'
+  if fp.exists():
+    print(f'>> ignore due to file exist: {fp}')
+    return
+
   real  = np.concatenate([it.real      for it in features]).flatten()
   imag  = np.concatenate([it.imag      for it in features]).flatten()
   mag   = np.concatenate([np.abs  (it) for it in features]).flatten()
@@ -51,7 +58,22 @@ def plot_hist(features:List[ndarray], title:str='feat'):
   plt.subplot(223) ; plt.hist(mag,   bins=300) ; plt.title('mag')
   plt.subplot(224) ; plt.hist(phase, bins=300) ; plt.title('phase')
   plt.suptitle(title)
+  print(f'>> save to {fp}')
+  plt.savefig(fp, dpi=600)
+  plt.close()
+
+
+def plot_scatter(features:List[ndarray], color:ndarray, title:str='feat'):
   fp = IMG_PATH / f'{title}.png'
+  if fp.exists():
+    print(f'>> ignore due to file exist: {fp}')
+    return
+
+  real = np.concatenate([it.real for it in features]).flatten()
+  imag = np.concatenate([it.imag for it in features]).flatten()
+
+  plt.scatter(real, imag, s=1, alpha=0.75, c=color, cmap='Spectral')
+  plt.suptitle(title)
   print(f'>> save to {fp}')
   plt.savefig(fp, dpi=600)
   plt.close()
@@ -59,23 +81,38 @@ def plot_hist(features:List[ndarray], title:str='feat'):
 
 if __name__ == '__main__':
   dataset = get_dataset()
-
   y = [it.y for it in dataset]
   plot_hist(y, title='y')
 
   for Nr in [64, 128]:
     y = [it.y for it in dataset if it.Nr == Nr]
+    color = np.concatenate([bits_to_number(it.bits) for it in dataset if it.Nr == Nr])
     plot_hist(y, title=f'y-Nr={Nr}')
+    plot_scatter(y, color, title=f'y-Nr={Nr}_scatter')
   for SNR in [10, 15, 20]:
     y = [it.y for it in dataset if it.SNR == SNR]
+    color = np.concatenate([bits_to_number(it.bits) for it in dataset if it.SNR == SNR])
     plot_hist(y, title=f'y-SNR={SNR}')
+    plot_scatter(y, color, title=f'y-SNR={SNR}_scatter')
 
   x = [it.x for it in dataset]
   plot_hist(x, title='x')
 
   for M in [4, 6, 8]:
     x = [it.x for it in dataset if it.num_bits_per_symbol == M]
+    color = np.concatenate([bits_to_number(it.bits) for it in dataset if it.num_bits_per_symbol == M])
     plot_hist(x, title=f'x-M={M}')
+    plot_scatter(x, color, title=f'x-M={M}_scatter')
   for SNR in [10, 15, 20]:
-    y = [it.x for it in dataset if it.SNR == SNR]
-    plot_hist(y, title=f'x-SNR={SNR}')
+    x = [it.x for it in dataset if it.SNR == SNR]
+    color = np.concatenate([bits_to_number(it.bits) for it in dataset if it.SNR == SNR])
+    plot_hist(x, title=f'x-SNR={SNR}')
+    plot_scatter(x, color, title=f'x-SNR={SNR}_scatter')
+
+  # sample-wise
+  for i, it in enumerate(dataset):
+    color = bits_to_number(it.bits)
+    plt.subplot(121) ; plt.scatter(it.y.real, it.y.imag, c=color, cmap='Spectral') ; plt.title('y')
+    plt.subplot(122) ; plt.scatter(it.x.real, it.x.imag, c=color, cmap='Spectral') ; plt.title('x')
+    plt.suptitle(f'id={i} SNR={it.SNR} nbps={it.num_bits_per_symbol}')
+    plt.show()
